requirements.txt
fastapi

google-generativeai
python-multipart
pillow
python-dotenv
aiofiles
duckduckgo-search

